+++ app/Http/Controllers/Admin/DepartmentController.php
@@ -8,6 +8,8 @@ use App\Models\Branch;
 use App\Models\Department;
 use Illuminate\Http\Request;
 use Exception;
+use Illuminate\Support\Facades\Gate;
+
 
 class DepartmentController extends Controller
 {
@@ -16,6 +18,8 @@ class DepartmentController extends Controller
      */
     public function index()
     {
+        abort_unless(Gate::allows('view department'), 403);
+
         $departments = Department::oldest('order')->paginate(perPage: 20);
         return view('admin.department.index', compact('departments'));
     }
@@ -25,6 +29,8 @@ class DepartmentController extends Controller
      */
     public function create()
     {
+        abort_unless(Gate::allows('create department'), 403);
+
         $branches = Branch::where('status', 1)->latest()->get();
         return view('admin.department.create', compact('branches'));
     }
@@ -34,6 +40,8 @@ class DepartmentController extends Controller
      */
     public function store(DepartmentRequest $request)
     {
+        abort_unless(Gate::allows('create department'), 403);
+
         try {
             $input = $request->all();
             $input['holidays'] = json_encode($request->holidays ?? []);
@@ -57,6 +65,8 @@ class DepartmentController extends Controller
      */
     public function edit(Department $department)
     {
+        abort_unless(Gate::allows('edit department'), 403);
+
         $branches = Branch::where('status', 1)->latest()->get();
         return view('admin.department.edit', compact('department', 'branches'));
     }
@@ -66,6 +76,8 @@ class DepartmentController extends Controller
      */
     public function update(DepartmentRequest $request, Department $department)
     {
+        abort_unless(Gate::allows('edit department'), 403);
+
         try {
             $input = $request->all();
             $input['holidays'] = $request->holidays ? json_encode($request->holidays) : json_encode([]);
@@ -81,6 +93,8 @@ class DepartmentController extends Controller
      */
     public function destroy(Department $department)
     {
+        abort_unless(Gate::allows('delete department'), 403);
+
         $department->delete();
         return redirect()->route('departments.index')->with('message', 'Delete Successfully');
     }
