+++ app/Http/Controllers/Admin/DesignationController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\Admin\DesignationRequest;
 use App\Models\Designation;
 use Illuminate\Http\Request;
 use Exception;
+use Illuminate\Support\Facades\Gate;
+
 
 class DesignationController extends Controller
 {
@@ -15,6 +17,8 @@ class DesignationController extends Controller
      */
     public function index()
     {
+        abort_unless(Gate::allows('view designation'), 403);
+
         $designations = Designation::oldest('order')->paginate(perPage: 20);
         return view('admin.designation.index', compact('designations'));
     }
@@ -24,6 +28,8 @@ class DesignationController extends Controller
      */
     public function create()
     {
+        abort_unless(Gate::allows('create designation'), 403);
+
         return view('admin.designation.create');
     }
 
@@ -32,6 +38,8 @@ class DesignationController extends Controller
      */
     public function store(DesignationRequest $request)
     {
+        abort_unless(Gate::allows('create designation'), 403);
+
         try {
             Designation::create($request->all());
             return redirect()->route('designations.index')->with('message', 'Designation Created Successfully');
@@ -53,6 +61,8 @@ class DesignationController extends Controller
      */
     public function edit(Designation $designation)
     {
+        abort_unless(Gate::allows('edit designation'), 403);
+
         return view('admin.designation.edit', compact('designation'));
     }
 
@@ -61,6 +71,8 @@ class DesignationController extends Controller
      */
     public function update(DesignationRequest $request, Designation $designation)
     {
+        abort_unless(Gate::allows('edit designation'), 403);
+
         try {
             $designation->update($request->all());
             return redirect()->route('designations.index')->with('message', 'Update Successfully');
@@ -74,6 +86,8 @@ class DesignationController extends Controller
      */
     public function destroy(Designation $designation)
     {
+        abort_unless(Gate::allows('delete designation'), 403);
+
         $designation->delete();
         return redirect()->route('designations.index')->with('message', 'Delete Successfully');
     }
