+++ app/Http/Controllers/Admin/LeavetypeController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Admin\LeavetypeRequest;
 use App\Models\LeaveType;
 use Illuminate\Http\Request;
 use Exception;
+use Illuminate\Support\Facades\Gate;
 
 class LeavetypeController extends Controller
 {
@@ -15,6 +16,8 @@ class LeavetypeController extends Controller
      */
     public function index()
     {
+        abort_unless(Gate::allows('view leavetype'), 403);
+
         $leavetypes = LeaveType::oldest('order')->paginate(perPage: 20);
         return view('admin.leavetype.index', compact('leavetypes'));
     }
@@ -24,6 +27,8 @@ class LeavetypeController extends Controller
      */
     public function create()
     {
+        abort_unless(Gate::allows('create leavetype'), 403);
+
         return view('admin.leavetype.create');
     }
 
@@ -32,6 +37,8 @@ class LeavetypeController extends Controller
      */
     public function store(LeavetypeRequest $request)
     {
+        abort_unless(Gate::allows('create leavetype'), 403);
+
         try {
             LeaveType::create($request->all());
             return redirect()->route('leavetypes.index')->with('message', 'Leavetype Created Successfully');
@@ -53,6 +60,8 @@ class LeavetypeController extends Controller
      */
     public function edit(LeaveType $leavetype)
     {
+        abort_unless(Gate::allows('edit leavetype'), 403);
+
         return view('admin.leavetype.edit', compact('leavetype'));
     }
 
@@ -61,6 +70,8 @@ class LeavetypeController extends Controller
      */
     public function update(LeavetypeRequest $request, LeaveType $leavetype)
     {
+        abort_unless(Gate::allows('edit leavetype'), 403);
+
         try {
             $input = $request->all();
             $input['is_paid'] = $request->is_paid ? 1 : 0;
@@ -76,6 +87,8 @@ class LeavetypeController extends Controller
      */
     public function destroy(LeaveType $leavetype)
     {
+        abort_unless(Gate::allows('delete leavetype'), 403);
+
         $leavetype->delete();
         return redirect()->route('leavetypes.index')->with('message', 'Delete Successfully');
     }
