+++ app/Http/Controllers/Admin/PublicHolidayController.php
@@ -8,6 +8,9 @@ use App\Models\Department;
 use App\Models\PublicHoliday;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Exception;
+
 
 class PublicHolidayController extends Controller
 {
@@ -16,6 +19,8 @@ class PublicHolidayController extends Controller
      */
     public function index()
     {
+        abort_unless(Gate::allows('view publicholiday'), 403);
+
         $publicHolidays = PublicHoliday::latest()->paginate(10);
         return view('admin.publicholiday.index', compact('publicHolidays'));
     }
@@ -25,6 +30,8 @@ class PublicHolidayController extends Controller
      */
     public function create()
     {
+        abort_unless(Gate::allows('create publicholiday'), 403);
+
         $departments = Department::orderBy('name', 'ASC')->get();
         return view('admin.publicholiday.create', compact('departments'));
     }
@@ -34,18 +41,23 @@ class PublicHolidayController extends Controller
      */
     public function store(PublicHolidayRequest $request)
     {
-        $input = $request->except('holidays');
-        $end_date = $request->end_date ?? $request->start_date;
-        $input['status'] = 1;
-        $input['end_date'] = $end_date;
+        abort_unless(Gate::allows('create publicholiday'), 403);
+        try {
+            $input = $request->except('holidays');
+            $end_date = $request->end_date ?? $request->start_date;
+            $input['status'] = 1;
+            $input['end_date'] = $end_date;
 
-        $start_date = Carbon::parse($request->start_date);
-        $end_date = Carbon::parse($end_date);
+            $start_date = Carbon::parse($request->start_date);
+            $end_date = Carbon::parse($end_date);
 
-        $input['total_days'] = $start_date->diffInDays($end_date) + 1;
-        $publicHoliday = PublicHoliday::create($input);
-        $publicHoliday->departments()->attach($request->departments);
-        return redirect()->route('publicholidays.index')->with('message', 'New Holiday Added');
+            $input['total_days'] = $start_date->diffInDays($end_date) + 1;
+            $publicHoliday = PublicHoliday::create($input);
+            $publicHoliday->departments()->attach($request->departments);
+            return redirect()->route('publicholidays.index')->with('message', 'New Holiday Added');
+        } catch (Exception $e) {
+            return redirect()->route('publicholidays.index')->with('warning', $e->getMessage())->withInput();
+        }
     }
 
     /**
@@ -61,6 +73,8 @@ class PublicHolidayController extends Controller
      */
     public function edit(PublicHoliday $publicholiday)
     {
+        abort_unless(Gate::allows('edit publicholiday'), 403);
+
         $departments = Department::orderBy('name', 'ASC')->get();
         return view('admin.publicholiday.edit', compact('publicholiday', 'departments'));
     }
@@ -70,19 +84,25 @@ class PublicHolidayController extends Controller
      */
     public function update(PublicHolidayRequest $request, PublicHoliday $publicholiday)
     {
-        $input = $request->except('holidays');
-        $end_date = $request->end_date ?? $request->start_date;
-        $input['status'] = 1;
-        $input['end_date'] = $end_date;
+        abort_unless(Gate::allows('edit publicholiday'), 403);
 
-        $start_date = Carbon::parse($request->start_date);
-        $end_date = Carbon::parse($end_date);
+        try {
+            $input = $request->except('holidays');
+            $end_date = $request->end_date ?? $request->start_date;
+            $input['status'] = 1;
+            $input['end_date'] = $end_date;
 
-        $input['total_days'] = $start_date->diffInDays($end_date) + 1;
-        $publicholiday->update($input);
-        $publicholiday->departments()->detach();
-        $publicholiday->departments()->attach($request->departments);
-        return redirect()->route('publicholidays.index')->with('message', 'Holiday Updated');
+            $start_date = Carbon::parse($request->start_date);
+            $end_date = Carbon::parse($end_date);
+
+            $input['total_days'] = $start_date->diffInDays($end_date) + 1;
+            $publicholiday->update($input);
+            $publicholiday->departments()->detach();
+            $publicholiday->departments()->attach($request->departments);
+            return redirect()->route('publicholidays.index')->with('message', 'Holiday Updated');
+        } catch (Exception $e) {
+            return redirect()->route('publicholidays.index')->with('warning', $e->getMessage())->withInput();
+        }
     }
 
     /**
@@ -90,6 +110,8 @@ class PublicHolidayController extends Controller
      */
     public function destroy(PublicHoliday $publicholiday)
     {
+        abort_unless(Gate::allows('delete publicholiday'), 403);
+
         $publicholiday->departments()->detach();
         $publicholiday->delete();
         return redirect()->route('publicholidays.index')->with('message', 'Holiday Deleted');
